+++ packages/sdk-next/shim.src.js
@@ -1,0 +1,9 @@
+import buffer from 'buffer';
+
+/**
+ * Fixes buffer issues
+ */
+if (typeof window !== 'undefined') {
+  // Code is running in the browser (client-side)
+  window.Buffer = buffer.Buffer;
+}