+++ packages/sdk-next/src/evm/index.ts
@@ -1,3 +1,4 @@
+export * from './adapter';
 export * from './chain';
 export * from './client';