+++ packages/sdk/src/aave/AaveClient.ts
@@ -1,5 +1,3 @@
-import { PublicClient } from 'viem';
-
 import { AAVE_POOL_ABI, AAVE_POOL_DATA_PROVIDER_ABI } from './abi';
 import {
   AAVE_LENDING_POOL_ADDRESS,
@@ -7,46 +5,46 @@ import {
   AAVE_POOL_PROXY,
 } from './const';
 
-import { EvmClient } from '../evm';
+import { EvmClient, EvmRpcAdapter } from '../evm';
 
 export class AaveClient {
-  private client: PublicClient;
+  private rpcAdapter: EvmRpcAdapter;
 
   constructor(evm: EvmClient) {
-    this.client = evm.getWsProvider();
+    this.rpcAdapter = evm.getRPCAdapter();
   }
 
   async getBlockTimestamp() {
-    const block = await this.client.getBlock();
+    const block = await this.rpcAdapter.getBlock();
     return Number(block.timestamp);
   }
 
   async getReservesData() {
-    const output = await this.client.readContract({
+    const output = await this.rpcAdapter.readContract({
       abi: AAVE_POOL_DATA_PROVIDER_ABI,
       address: AAVE_POOL_DATA_PROVIDER as `0x${string}`,
-      args: [AAVE_LENDING_POOL_ADDRESS as `0x${string}`],
       functionName: 'getReservesData',
+      args: [AAVE_LENDING_POOL_ADDRESS as `0x${string}`],
     });
     return output;
   }
 
   async getUserReservesData(user: string) {
-    const output = await this.client.readContract({
+    const output = await this.rpcAdapter.readContract({
       abi: AAVE_POOL_DATA_PROVIDER_ABI,
       address: AAVE_POOL_DATA_PROVIDER as `0x${string}`,
-      args: [AAVE_LENDING_POOL_ADDRESS as `0x${string}`, user as `0x${string}`],
       functionName: 'getUserReservesData',
+      args: [AAVE_LENDING_POOL_ADDRESS as `0x${string}`, user as `0x${string}`],
     });
     return output;
   }
 
   async getUserAccountData(user: string) {
-    const output = await this.client.readContract({
+    const output = await this.rpcAdapter.readContract({
       abi: AAVE_POOL_ABI,
       address: AAVE_POOL_PROXY as `0x${string}`,
-      args: [user as `0x${string}`],
       functionName: 'getUserAccountData',
+      args: [user as `0x${string}`],
     });
     return output;
   }