+++ packages/sdk/src/tx/TradeTxBuilder.ts
@@ -77,7 +77,7 @@ export class TradeTxBuilder extends TxBuilder {
     return this.buildSellTx();
   }
 
-  private buildBuyTx(): SubstrateTransaction {
+  private async buildBuyTx(): Promise<SubstrateTransaction> {
     const { amountIn, amountOut, swaps } = this.trade;
 
     const firstSwap = swaps[0];
@@ -106,6 +106,12 @@ export class TradeTxBuilder extends TxBuilder {
         maxAmountIn.toFixed(),
         TradeRouteBuilder.build(swaps)
       );
+    }
+
+    const hasDebt = await this.aaveUtils.hasBorrowPositions(this.beneficiary);
+    if (hasDebt) {
+      const txWithExtraGas = this.dispatchWithExtraGas(tx);
+      return this.wrapTx('RouterBuy', txWithExtraGas, AAVE_EXTRA_GAS);
     }
 
     return this.wrapTx('RouterBuy', tx);