+++ packages/xcm-cfg/src/configs/polkadot/astar.ts
@@ -35,28 +35,28 @@ const toHydration: AssetRoute[] = [
     },
     extrinsic: ExtrinsicBuilder().polkadotXcm().reserveTransferAssets(),
   }),
-  new AssetRoute({
-    source: {
-      asset: dot,
-      balance: BalanceBuilder().substrate().assets().account(),
-      fee: {
-        asset: astr,
-        balance: BalanceBuilder().substrate().system().account(),
-      },
-      destinationFee: {
-        balance: BalanceBuilder().substrate().assets().account(),
-      },
-    },
-    destination: {
-      chain: hydration,
-      asset: dot,
-      fee: {
-        amount: 0.1,
-        asset: dot,
-      },
-    },
-    extrinsic: ExtrinsicBuilder().xTokens().transfer(),
-  }),
+  // new AssetRoute({
+  //   source: {
+  //     asset: dot,
+  //     balance: BalanceBuilder().substrate().assets().account(),
+  //     fee: {
+  //       asset: astr,
+  //       balance: BalanceBuilder().substrate().system().account(),
+  //     },
+  //     destinationFee: {
+  //       balance: BalanceBuilder().substrate().assets().account(),
+  //     },
+  //   },
+  //   destination: {
+  //     chain: hydration,
+  //     asset: dot,
+  //     fee: {
+  //       amount: 0.1,
+  //       asset: dot,
+  //     },
+  //   },
+  //   extrinsic: ExtrinsicBuilder().xTokens().transfer(),
+  // }),
   new AssetRoute({
     source: {
       asset: bnc,