+++ packages/xcm-cfg/src/configs/polkadot/hydration/index.ts
@@ -203,7 +203,7 @@ const toAssetHub: AssetRoute[] = [
 
 const toAstar: AssetRoute[] = [
   toTransferTemplate(astr, astar, 0.00404146544),
-  toTransferTemplate(dot, astar, 0.1),
+  //toTransferTemplate(dot, astar, 0.1),
   toTransferTemplate(bnc, astar, 0.75),
   toTransferTemplate(glmr, astar, 0.025),
   toTransferTemplate(ibtc, astar, 0.000002),
@@ -219,7 +219,7 @@ const toBifrost: AssetRoute[] = [
   toTransferTemplate(bnc, bifrost, 0.000563136),
   toTransferTemplate(vdot, bifrost, 0.0000000703),
   toTransferTemplate(vastr, bifrost, 0.000000643),
-  toTransferTemplate(dot, bifrost, 0.1),
+  //toTransferTemplate(dot, bifrost, 0.1),
   toTransferTemplate(astr, bifrost, 0.5),
   toTransferTemplate(glmr, bifrost, 0.1),
   toTransferTemplate(ibtc, bifrost, 0.000005),
@@ -229,14 +229,14 @@ const toBifrost: AssetRoute[] = [
 
 const toCentrifuge: AssetRoute[] = [
   toTransferTemplate(cfg, centrifuge, 0.0092696),
-  toTransferTemplate(dot, centrifuge, 0.1),
+  //toTransferTemplate(dot, centrifuge, 0.1),
   toTransferTemplate(glmr, centrifuge, 0.05),
 ];
 
 const toInterlay: AssetRoute[] = [
   toTransferTemplate(ibtc, interlay, 0.00000062),
   toTransferTemplate(intr, interlay, 0.0019213457),
-  toTransferTemplate(dot, interlay, 0.1),
+  //toTransferTemplate(dot, interlay, 0.1),
   toTransferTemplate(hdx, interlay, 0.5),
   toTransferTemplate(usdt, interlay, 0.3),
   toTransferTemplate(usdc, interlay, 0.3),
@@ -264,7 +264,7 @@ const toInterlay: AssetRoute[] = [
 const toMoonbeam: AssetRoute[] = [
   toTransferTemplate(hdx, moonbeam, 5),
   toTransferTemplate(glmr, moonbeam, 0.01),
-  toTransferTemplate(dot, moonbeam, 0.1),
+  //toTransferTemplate(dot, moonbeam, 0.1),
   toTransferTemplate(usdt, moonbeam, 0.3),
   toTransferTemplate(usdc, moonbeam, 0.3),
   toMoonbeamErc20Template(dai_mwh),