+++ packages/xcm-cfg/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @galacticcouncil/xcm-cfg
 
+## 10.15.3
+
+### Patch Changes
+
+- reenable snowfork
+
+## 10.15.2
+
+### Patch Changes
+
+- temp disable dot transfers from paras -> hydration
+
+## 10.15.1
+
+### Patch Changes
+
+- dfe64d1: cfg: enable snowbridge swap
+
 ## 10.15.0
 
 ### Minor Changes